+++ firmware/spi.hh
@@ -28,7 +28,7 @@ bool load_bitstream_flash(std::uint8_t slot_index) noexcept;
 bool load_bitstream_psram() noexcept;
 /* Shuffle the data from the PSRAM to FLASH, note that expected_len does not include the header size */
 [[nodiscard]]
-bool move_to_slot(std::uint8_t slot_index, std::uint16_t expected_len) noexcept;
+bool move_to_slot(std::uint8_t slot_index, std::uint32_t expected_len) noexcept;
 
 [[nodiscard]]
 std::uint8_t read_squishy_register(std::uint8_t addr) noexcept;
